Created .gitignore, and force.d file. Included in main, fixed Makefile
Made a .gitignore file to iginore the files in the doc, bin, obj dirs. Created
base of force.d file which starts the force class. Made main.d included that so
we can make sure it compiles. Mkakefile is fixed so that it automatically
cleans and makes the directories on the default target.