Created interface class entity
Created the interface entity in entity.d, will be used as supr class for
all thing that will be drawn to screen. Made display.d have a list of
Entities and a way to add them. Its Draw iterates through them and
Draw's each of them. DevIL was taken out of Display. ProcessInput was
added to trap events such as key presses for now, it only checks to see
if the window is active or not, and if it gets the close signal. main.d
has beeen updated to take into account displays new functionality, and
enters an infinite loop of calling various display functions.