Created the point class
Point is a class implementing the Entity specification. It has several
constructors that range in the amount of variables that are predefined,
the first defining only X and Y, the last defining everything. Draw
takes into account position, color, and size. There are {g,s}etters for
each aspect of the color and radius, and two special Color setters for
RGB and RGBA specification.