Bunch of udpates for forces.
Everything force wise works so far, but the interval of update should be
split from the rendering engine, as this causes problems.

Created and semi-implemented gravity/repel forces. Gravity was supposed
to do everything, but the update interval is not the best. Repel was
supposed to prevent collisions/intersections from happening by having a
large repel at close range. That failed. If the update interval was
abstracted from rendering, that would be best. They're too tied
together.

Parcticle was updated to hold multiple entities/forces. Forces can be
worked on other things. Lots of new functions for dealing with this.

Particle system was created, and appears to work. Only problem is delta
time should be taken into account with the working of forces. Huge todo
right there. Pretty much eveyrthing else works though.