Fixed up repel quite a bit
Made repel work properly on both B and A particles, as there were
numerous typos. Changed the way it works so it adds a counteractive
force to it's current vector of closeness.

Made main.d take advntage of some new stuff, mainly by creating a 5th
particle that moves in a circle. Used the active/nonactive hook as a way
to reset the particles once they go all crazy.