Bunches of changes to try and get deltaTime to work
Gravity has been overhauled to only affect one particle at a time, and
not to work if the objects are overlapping, so it simulates a point mass
that once something gets to close to it the mass is inside the object.
Gravity know also has a gravity constant, which is quite a bit different
the the supposed real one.

Repel has had some work done, trying to figure out how to calculate the
momentum change. To help with this, particle has been changed so that it
buffers changes until it is updated. This is so everything can have work
done on it, and then have that work actually applied.

Main.d has been altered to accomidate these changes, and has added a
large stationary mass at the center of the screen.

Particle has also had the acceleration be "used up" so that once it has
been calculated, it is zero'd.

Particle system has been overhauled again to have a few usefull
functions, a step, and a deltaTime passed to it. The internal deltaTime
varies according to what is needed to have the fastest partcile moving
at step, with the sum of all elapsed time equalling the argument.