Various new updates: see commit message
Added new stuff to TODO. Remembered about TODO tags in vim, so now there are two of those in the
same place ^^. Point now has a zoom factor and draws it's radius accordingly. To go with that, game
has had some mouswheel hacks because of the way it spawns two events for the same thing. Seems to
work. main takes advantage by allowing zooming in/out, but it only looks normal in lock mode because
it does no coordinate adjustion. Main now also allows for dynamic particle creation, which needs to
have a better algorithm (read: remotely correct) for mapping screen space to actual space
coordinates. Other than that, it looks good.