Started moving towards a way to delete particles
Mouse tracking has been set up. Basically, the only thing that needs to be done now is to have the
particles all in a linked list of some kind, so we can delete ones out of the middle. Also, a way to
remove the corresponding particle should probably be set up. Like have a pointer to particle in each
point, and then go through them all and if the have no corresponding particle just delete them.