Changed class based Force's into funtion pointer based templates
Force is now a typedef of a particular type of function pointer that should work exactly as the old
Work function was used. It makes sure that everything returns void and accepts the proper arguments.
All of the forces are now template functions. They work exactly the same, except now they are bare
functions (instead of being wrapped in a class). They are templates so that you can have different
constants of effect, as before. Best to look in here and see what is going on. A Default<Force> is
made to have the compiler check to make sure it conforms to the Force typedef.
Main has been modified slightly to now make Force m_<Blah> instead of <Blah> m_<Blah> which point to
a specific specilization of <blah>'s template function.
Particle had one change, instead of calling Work on each Force object, it now calls the Force
function pointer directly.